+++ branches/trunk/GUI/spot-on.h
@@ -514,7 +514,8 @@ class spoton: public QMainWindow
   void demagnetize(void);
   void derivativeUpdates(void);
   void discoverUrls(void);
-  void displayUrlImportResults(const int imported, const int not_imported);
+  void displayUrlImportResults(const quint64 imported,
+			       const quint64 not_imported);
   void generateHalfGeminis(void);
   void highlightPaths(void);
   void importNeighbors(const QString &filePath);
